refactor: sprints 42.3 + 42.4 — scancodeio dtos.py + endpoints.py
Completes the scancodeio_client.py split started in commit ede3490
(Sprint 42.1+42.2). Audit item 1.2 closed.

- libinv/services/scancodeio/dtos.py (new): TypedDict-style payload
  schemas previously inlined in __init__.py.
- libinv/services/scancodeio/endpoints.py (new): every endpoint-handler
  method (list_projects, get_project, create_project, ...) moved here
  as `class EndpointsMixin`. `ScancodeioClient(EndpointsMixin)` in
  __init__.py is now a thin façade composing Transport (Sprint 42.2)
  with EndpointsMixin.
- The CHANGELOG-flagged `list_projects_for_wasp` + `compare_builds.py`
  blocker (upstream `ProjectFilterSet.Meta.fields` extension needed)
  is now documented inline in endpoints.py near the affected method.
  Method retained, not deleted.

Verification: 20/20 scancodeio tests pass; mypy clean on the 4-file
package; integration baseline preserved (90/3/1).
